s3 は特別扱いする
