+++ s3/s3_test.go
@@ -1,4 +1,4 @@
-package archive
+package s3
 
 import (
 	"testing"
